#15 Separate auth related files to /auth/ subdirectories and user table files to /models/ subdirectories
