+++ backend/src/database/crud/user_crud.py
@@ -1,7 +1,7 @@
 from backend.src.core.security import hash_password
 from sqlalchemy.orm.session import Session
 from backend.src.database.models.user_model import User
-from backend.src.schemas.user_schema import UserCreate
+from backend.src.schemas.model.user_schema import UserCreate
 
 def get_user(db: Session, user_id: int) -> type[User] | None:
     return db.query(User).filter(User.id == user_id).first()
