+++ backend/src/routes/auth/auth_router.py
@@ -4,8 +4,8 @@ from typing import Annotated
 from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session
 from backend.src.database.session import get_db
-from backend.src.schemas.token_schema import Token
-from backend.src.services.auth_service import authenticate_user, create_access_token
+from backend.src.schemas.auth.token_schema import Token
+from backend.src.services.auth.auth_service import authenticate_user, create_access_token
 
 auth_router = APIRouter(
     prefix="/auth",
